Add a hack fix to make quakenet work
Quakenet refuses to complete registration until you PONG to their first PING.
We had previously been deferring the PONG until registration is complete.

This is somewhat of a hack on top of a pile of hacks - the entire registration send process needs fixing.
We probably want the ChunkedPriorityQueue implementation from my misc libs.