client: Fix a race condition around nick changing where we desync from what server thinks our nick is
A bug occurred where wait_for_messages() timed out while changing nick.
This led us to erroniously accept the new nick as our nick. Then later,
when a NICKNAMEINUSE arrived, we discarded it as we were not currently changing nicks,
leaving us in a situation where what we thought our nick was, was not what the server thought.

We implement three mitigations here:
* Increase the wait_for_messages timeout to try to mitigate network blips
* Add a LOT more logging around nick changes and potential pitfalls to make debugging easier
* If this scenario occurs again, we react to the subsequent NICKINUSE by assuming we fucked up
  and returning to whatever the server tells us is currently our nick.